% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-infrastructure.R
\name{get_rtistic_palettes}
\alias{get_rtistic_palettes}
\title{Get names of all unique palettes provided in Rtistic}
\usage{
get_rtistic_palettes()
}
\value{
Vector of palette name stubs
}
\description{
Get names of all unique palettes provided in Rtistic
}
\examples{
get_rtistic_palettes()
}
