% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtistic.R
\docType{package}
\name{Rtistic}
\alias{Rtistic}
\alias{Rtistic-package}
\title{\code{Rtistic} package}
\description{
A collection of palettes and themes for plots, documents, and slides
}
\details{
See the README on
\href{https://github.com/emilyriederer/Rtistic}{GitHub}
}
