% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-infrastructure.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_discrete_rtistic}
\alias{scale_colour_discrete_rtistic}
\alias{scale_fill_discrete_rtistic}
\alias{scale_colour_continuous_div_rtistic}
\alias{scale_color_continuous_div_rtistic}
\alias{scale_fill_continuous_div_rtistic}
\alias{scale_colour_continuous_rtistic}
\alias{scale_color_continuous_rtistic}
\alias{scale_fill_continuous_rtistic}
\title{Custom Color and Fill Scales}
\usage{
scale_color_discrete_rtistic(palette = "test", extend = FALSE, ...)

scale_colour_discrete_rtistic(palette = "test", extend = FALSE, ...)

scale_fill_discrete_rtistic(palette = "test", extend = FALSE, ...)

scale_colour_continuous_div_rtistic(palette = "test", ...)

scale_color_continuous_div_rtistic(palette = "test", ...)

scale_fill_continuous_div_rtistic(palette = "test", ...)

scale_colour_continuous_rtistic(palette = "test", ...)

scale_color_continuous_rtistic(palette = "test", ...)

scale_fill_continuous_rtistic(palette = "test", ...)
}
\description{
Custom coloring and filling functions based on unique palettes
}
\references{
https://en.wikipedia.org/wiki/Traffic_light
}
