% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-slide-test.R
\name{slides_test}
\alias{slides_test}
\title{Test HTML template}
\usage{
slides_test(...)
}
\arguments{
\item{...}{additional arguments provided to \code{moon_reader}}
}
\description{
Loads additional style and template file
}
\references{
https://raw.githubusercontent.com/dr-harper/example-rmd-templates/master/R/my_html_format.R
}
