% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-palette-test.R
\docType{data}
\name{test_pal}
\alias{test_pal}
\alias{test_pal_op}
\alias{test_pal_cont}
\alias{test_pal_div}
\title{Test palette}
\format{An object of class \code{character} of length 4.}
\usage{
test_pal

test_pal_op

test_pal_cont

test_pal_div
}
\description{
This is a test palette inspired by stop-light colors
}
\references{
https://en.wikipedia.org/wiki/Traffic_light
}
\keyword{datasets}
