% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-palette.R
\name{get_rtistic_palettes}
\alias{get_rtistic_palettes}
\title{Get names of all unique palettes provided in Rtistic}
\usage{
get_rtistic_palettes(full = FALSE)
}
\arguments{
\item{full}{Whether to include full palette names (with suffixes, e.g. \code{_cont}) or just stubs}
}
\value{
Vector of palette name stubs or full names
}
\description{
Get names of all unique palettes provided in Rtistic
}
\examples{
get_rtistic_palettes()
}
