% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-gg-theme.R
\name{my_theme}
\alias{my_theme}
\title{Test ggplot theme}
\usage{
my_theme(base_theme = ggplot2::theme_minimal())
}
\arguments{
\item{base_theme}{Starting theme of plot, default is
[ggplot2::theme_minimal()]. Any elements set by `theme_pomological()` will
overwrite the `base_theme` unless the specific parameter is explicitly set
to `NULL`.}
}
\description{
[ggplot2] plot theme based on...
}
\examples{
\dontrun{
library(ggplot2)
data <- data.frame(x = 1:10, y = 1:10)
ggplot(data, aes(x, y)) + geom_point() + my_theme()
}

}
\references{
https://github.com/gadenbuie/ggpomological/blob/master/R/theme_pomological.R
}
\seealso{
[ggplot2::theme]
}
