% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-palette.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_discrete_rtistic}
\alias{scale_colour_discrete_rtistic}
\alias{scale_fill_discrete_rtistic}
\alias{scale_color_opinionated_rtistic}
\alias{scale_colour_opinionated_rtistic}
\alias{scale_fill_opinionated_rtistic}
\alias{scale_color_diverging_rtistic}
\alias{scale_colour_diverging_rtistic}
\alias{scale_fill_diverging_rtistic}
\alias{scale_color_continuous_rtistic}
\alias{scale_colour_continuous_rtistic}
\alias{scale_fill_continuous_rtistic}
\title{Custom color and fill scales}
\usage{
scale_color_discrete_rtistic(palette = "test", extend = FALSE, ...)

scale_colour_discrete_rtistic(palette = "test", extend = FALSE, ...)

scale_fill_discrete_rtistic(palette = "test", extend = FALSE, ...)

scale_color_opinionated_rtistic(palette = "test", ...)

scale_colour_opinionated_rtistic(palette = "test", extend = FALSE, ...)

scale_fill_opinionated_rtistic(palette = "test", ...)

scale_color_diverging_rtistic(palette = "test", ...)

scale_colour_diverging_rtistic(palette = "test", ...)

scale_fill_diverging_rtistic(palette = "test", ...)

scale_color_continuous_rtistic(palette = "test", ...)

scale_colour_continuous_rtistic(palette = "test", ...)

scale_fill_continuous_rtistic(palette = "test", ...)
}
\arguments{
\item{palette}{Name of color palette}

\item{extend}{Whether to extend discrete color palette to make sufficient colors for levels needed}

\item{...}{Additional arguments to be passed to internal scale function}
}
\description{
Custom coloring and filling functions based on unique color palettes
}
\details{
Specific functions include:
\itemize{
\item{scale_(color/colour/fill)_discrete_rtistic}{Discrete palette with either fixed or dynamically extended number of shades}
\item{scale_(color/colour/fill)_opinionated_rtistic}{Discrete palette with specific values for "good", "bad", and "neutral"}
\item{scale_(color/colour/fill)_diverging_rtistic}{Continuous diverging color palette}
\item{scale_(color/colour/fill)_continuous}{COntinuous color palette}
}
}
\references{
https://en.wikipedia.org/wiki/Traffic_light
}
