% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-theme-test.R
\name{theme_test}
\alias{theme_test}
\title{Test ggplot theme}
\usage{
theme_test(base_theme = ggplot::theme_minimal())
}
\arguments{
\item{base_theme}{A base theme upon which additional theme-specific options are applied}
}
\description{
[ggplot2] plot theme based on...
}
\examples{
\dontrun{
library(ggplot2)
data <- data.frame(x = 1:10, y = 1:10)
ggplot(data, aes(x, y)) + geom_point() + theme_test()
}


}
\references{
https://github.com/gadenbuie/ggpomological/blob/master/R/theme_pomological.R
}
\seealso{
[ggplot2::theme]
}
